# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API文档中心",
        "contact": {
            "name": "代先华",
            "url": "hc1989.com",
            "email": "d536819@sina.cn"
        },
        "version": "3.0"
    },
    "servers": [
        {
            "url": "http://106.54.212.105:8088"
        }
    ],
    "paths": {
        "/api/v1/users": {
            "post": {
                "tags": [
                    "用户注册"
                ],
                "summary": "用户注册地址",
                "operationId": "App\\Http\\Controllers\\Api\\UserController::store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "用户名",
                        "required": true,
                        "allowEmptyValue": false,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "ddd"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "密码",
                        "required": true,
                        "allowEmptyValue": false,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "123456"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "响应成功"
                    },
                    "400": {
                        "description": "响应失败"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "获取token值"
                ],
                "summary": "返回用户登陆的token值",
                "operationId": "App\\Http\\Controllers\\Api\\UserController::login",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "登录名",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "xander13"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "登陆密码",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "123456"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "请求成功"
                    },
                    "400": {
                        "description": "请求失败"
                    }
                }
            }
        },
        "/api/v1/personalInfo": {
            "post": {
                "tags": [
                    "身份证实名认证"
                ],
                "summary": "身份证实名认证",
                "operationId": "App\\Http\\Controllers\\Api\\UserController::personalInfo",
                "parameters": [
                    {
                        "name": "idcard",
                        "in": "query",
                        "description": "身份证号码",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "513022199107121677"
                    },
                    {
                        "name": "realname",
                        "in": "query",
                        "description": "姓名",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "代先华"
                    }
                ],
                "responses": {
                    "205": {
                        "description": "请求成功"
                    },
                    "400": {
                        "description": "请求失败"
                    }
                }
            }
        },
...
    }
}